Added solve24function.wl
Wolfram Language code for the solve24 function. Accepts a list of four integers, returns a list of strings representing equations produced from those four integers and the four basic arithmetic operations addition, subtraction, division and multiplication in every order of precedence, the result of which is 24. Useful to solve puzzles in the https://www.24game.com card game.